+++ packages/firebase_auth/lib/firebase_auth.dart
@@ -38,6 +38,7 @@ class UserInfo {
 }
 
 /// Represents user profile data that can be updated by [updateProfile]
+///
 /// The purpose of having separate class with a map is to give possibility
 /// to check if value was set to null or not provided
 class UserUpdateInfo {
@@ -74,28 +75,54 @@ class FirebaseUser extends UserInfo {
 
   /// Obtains the id token for the current user, forcing a [refresh] if desired.
   ///
+  /// Useful when authenticating against your own backend. Use our server
+  /// SDKs or follow the official documentation to securely verify the
+  /// integrity and validity of this token.
+  ///
   /// Completes with an error if the user is signed out.
+  ///
   Future<String> getIdToken({bool refresh = false}) async {
     return await FirebaseAuth.channel.invokeMethod('getIdToken', <String, bool>{
       'refresh': refresh,
     });
   }
 
+  /// Initiates email verification for the user.
+  ///
   Future<void> sendEmailVerification() async {
     await FirebaseAuth.channel.invokeMethod('sendEmailVerification');
   }
 
-  /// Manually refreshes the data of the current user (for example, attached providers, display name, and so on).
+  /// Manually refreshes the data of the current user (for example,
+  /// attached providers, display name, and so on).
+  ///
   Future<void> reload() async {
     await FirebaseAuth.channel.invokeMethod('reload');
   }
 
   /// Deletes the user record from your Firebase project's database.
+  ///
   Future<void> delete() async {
     await FirebaseAuth.channel.invokeMethod('delete');
   }
 
   /// Updates the email address of the user.
+  ///
+  /// The original email address recipient will receive an email that allows
+  /// them to revoke the email address change, in order to protect them
+  /// from account hijacking.
+  ///
+  /// **Important**: This is a security sensitive operation that requires
+  /// the user to have recently signed in.
+  ///
+  /// Errors:
+  ///   • `ERROR_INVALID_CREDENTIAL` - If the email address is malformed.
+  ///   • `ERROR_EMAIL_ALREADY_IN_USE` - If the email is already in use by a different account.
+  ///   • `ERROR_USER_DISABLED` - If the user has been disabled (for example, in the Firebase console)
+  ///   • `ERROR_USER_NOT_FOUND` - If the user has been deleted (for example, in the Firebase console)
+  ///   • `ERROR_REQUIRES_RECENT_LOGIN` - If the user's last sign-in time does not meet the security threshold.
+  ///   • `ERROR_OPERATION_NOT_ALLOWED` - Indicates that Email & Password accounts are not enabled.
+  ///
   Future<void> updateEmail(String email) async {
     assert(email != null);
     return await FirebaseAuth.channel.invokeMethod(
@@ -105,6 +132,20 @@ class FirebaseUser extends UserInfo {
   }
 
   /// Updates the password of the user.
+  ///
+  /// Anonymous users who update both their email and password will no
+  /// longer be anonymous. They will be able to log in with these credentials.
+  ///
+  /// **Important**: This is a security sensitive operation that requires
+  /// the user to have recently signed in.
+  ///
+  /// Errors:
+  ///   • `ERROR_WEAK_PASSWORD` - If the password is not strong enough.
+  ///   • `ERROR_USER_DISABLED` - If the user has been disabled (for example, in the Firebase console)
+  ///   • `ERROR_USER_NOT_FOUND` - If the user has been deleted (for example, in the Firebase console)
+  ///   • `ERROR_REQUIRES_RECENT_LOGIN` - If the user's last sign-in time does not meet the security threshold.
+  ///   • `ERROR_OPERATION_NOT_ALLOWED` - Indicates that Email & Password accounts are not enabled.
+  ///
   Future<void> updatePassword(String password) async {
     assert(password != null);
     return await FirebaseAuth.channel.invokeMethod(
@@ -114,11 +155,92 @@ class FirebaseUser extends UserInfo {
   }
 
   /// Updates the user profile information.
+  ///
+  /// Errors:
+  ///   • `ERROR_USER_DISABLED` - If the user has been disabled (for example, in the Firebase console)
+  ///   • `ERROR_USER_NOT_FOUND` - If the user has been deleted (for example, in the Firebase console)
+  ///
   Future<void> updateProfile(UserUpdateInfo userUpdateInfo) async {
     assert(userUpdateInfo != null);
     return await FirebaseAuth.channel.invokeMethod(
       'updateProfile',
       userUpdateInfo._updateData,
+    );
+  }
+
+  /// Detaches Email & Password from this user.
+  ///
+  /// This detaches the Email & Password from the current user. This will
+  /// prevent the user from signing in to this account with those credentials.
+  ///
+  /// **Important**: This is a security sensitive operation that requires
+  /// the user to have recently signed in.
+  ///
+  /// Errors:
+  ///   • `ERROR_NO_SUCH_PROVIDER` - If the user does not have an Email & Password linked to their account.
+  ///   • `ERROR_REQUIRES_RECENT_LOGIN` - If the user's last sign-in time does not meet the security threshold.
+  ///
+  Future<void> unlinkEmailAndPassword() async {
+    return await FirebaseAuth.channel.invokeMethod(
+      'unlinkCredential',
+      const <String, String>{'provider': 'password'},
+    );
+  }
+
+  /// Detaches Google from this user.
+  ///
+  /// This detaches the Google Account from the current user. This will
+  /// prevent the user from signing in to this account with those credentials.
+  ///
+  /// **Important**: This is a security sensitive operation that requires
+  /// the user to have recently signed in.
+  ///
+  /// Errors:
+  ///   • `ERROR_NO_SUCH_PROVIDER` - If the user does not have a Google Account linked to their account.
+  ///   • `ERROR_REQUIRES_RECENT_LOGIN` - If the user's last sign-in time does not meet the security threshold.
+  ///
+  Future<void> unlinkGoogleCredential() async {
+    return await FirebaseAuth.channel.invokeMethod(
+      'unlinkCredential',
+      const <String, String>{'provider': 'google.com'},
+    );
+  }
+
+  /// Detaches Facebook from this user.
+  ///
+  /// This detaches the Facebook Account from the current user. This will
+  /// prevent the user from signing in to this account with those credentials.
+  ///
+  /// **Important**: This is a security sensitive operation that requires
+  /// the user to have recently signed in.
+  ///
+  /// Errors:
+  ///   • `ERROR_NO_SUCH_PROVIDER` - If the user does not have a Facebook Account linked to their account.
+  ///   • `ERROR_REQUIRES_RECENT_LOGIN` - If the user's last sign-in time does not meet the security threshold.
+  ///
+  Future<void> unlinkFacebookCredential() async {
+    return await FirebaseAuth.channel.invokeMethod(
+      'unlinkCredential',
+      const <String, String>{'provider': 'facebook.com'},
+    );
+  }
+
+  /// Detaches Twitter from this user.
+  ///
+  /// This detaches the Twitter Account from the current user. This will
+  /// prevent the user from signing in to this account with those credentials.
+  ///
+  /// **Important**: This is a security sensitive operation that requires
+  /// the user to have recently signed in.
+  ///
+  /// Errors:
+  ///   • `ERROR_NO_SUCH_PROVIDER` - If the user does not have a Twitter Account linked to their account.
+  ///   • `ERROR_REQUIRES_RECENT_LOGIN` - If the user's last sign-in time does not meet the security threshold.
+  ///
+  Future<void> unlinkTwitterCredential() async {
+    return await FirebaseAuth.channel.invokeMethod(
+      'unlinkCredential',
+      const <String, String>{'provider': 'twitter.com'},
     );
   }
 
@@ -161,7 +283,7 @@ class FirebaseAuth {
     channel.setMethodCallHandler(_callHandler);
   }
 
-  /// Receive [FirebaseUser] each time the user signIn or signOut
+  /// Receive [FirebaseUser] each time the user signIn or signOut or has changed.
   Stream<FirebaseUser> get onAuthStateChanged {
     Future<int> _handle;
 
@@ -190,9 +312,12 @@ class FirebaseAuth {
   /// returned instead. If there is any other existing user signed in, that
   /// user will be signed out.
   ///
-  /// Will throw a PlatformException if
-  /// FIRAuthErrorCodeOperationNotAllowed - Indicates that anonymous accounts are not enabled. Enable them in the Auth section of the Firebase console.
-  /// See FIRAuthErrors for a list of error codes that are common to all API methods.
+  /// **Important**: You must enable Anonymous accounts in the Auth section
+  /// of the Firebase console before being able to use them.
+  ///
+  /// Errors:
+  ///   • `ERROR_OPERATION_NOT_ALLOWED` - Indicates that Anonymous accounts are not enabled.
+  ///
   Future<FirebaseUser> signInAnonymously() async {
     final Map<dynamic, dynamic> data =
         await channel.invokeMethod('signInAnonymously');
@@ -200,6 +325,16 @@ class FirebaseAuth {
     return currentUser;
   }
 
+  /// Tries to create a new user account with the given email address and password.
+  ///
+  /// If successful, it also signs the user in into the app and updates
+  /// the [onAuthStateChanged] stream.
+  ///
+  /// Errors:
+  ///   • `ERROR_WEAK_PASSWORD` - If the password is not strong enough.
+  ///   • `ERROR_INVALID_CREDENTIAL` - If the email address is malformed.
+  ///   • `ERROR_EMAIL_ALREADY_IN_USE` - If the email is already in use by a different account.
+  ///
   Future<FirebaseUser> createUserWithEmailAndPassword({
     @required String email,
     @required String password,
@@ -217,12 +352,28 @@ class FirebaseAuth {
     return currentUser;
   }
 
+  @Deprecated("Please use fetchSignInMethodsForEmail() instead.")
   Future<List<String>> fetchProvidersForEmail({
+    @required String email,
+  }) =>
+      fetchSignInMethodsForEmail(email: email);
+
+  /// Returns a list of sign-in methods that can be used to sign in a given
+  /// user (identified by its main email address).
+  ///
+  /// This method is useful when you support multiple authentication mechanisms
+  /// if you want to implement an email-first authentication flow.
+  ///
+  /// Errors:
+  ///   • `ERROR_INVALID_CREDENTIAL` - If the [email] address is malformed.
+  ///   • `ERROR_USER_NOT_FOUND` - If there is no user corresponding to the given [email] address.
+  ///
+  Future<List<String>> fetchSignInMethodsForEmail({
     @required String email,
   }) async {
     assert(email != null);
     final List<dynamic> providers = await channel.invokeMethod(
-      'fetchProvidersForEmail',
+      'fetchSignInMethodsForEmail',
       <String, String>{
         'email': email,
       },
@@ -230,6 +381,14 @@ class FirebaseAuth {
     return providers?.cast<String>();
   }
 
+  /// Triggers the Firebase Authentication backend to send a password-reset
+  /// email to the given email address, which must correspond to an existing
+  /// user of your app.
+  ///
+  /// Errors:
+  ///   • `ERROR_INVALID_EMAIL` - If the [email] address is malformed.
+  ///   • `ERROR_USER_NOT_FOUND` - If there is no user corresponding to the given [email] address.
+  ///
   Future<void> sendPasswordResetEmail({
     @required String email,
   }) async {
@@ -242,6 +401,22 @@ class FirebaseAuth {
     );
   }
 
+  /// Tries to sign in a user with the given email address and password.
+  ///
+  /// If successful, it also signs the user in into the app and updates
+  /// the [onAuthStateChanged] stream.
+  ///
+  /// **Important**: You must enable Email & Password accounts in the Auth
+  /// section of the Firebase console before being able to use them.
+  ///
+  /// Errors:
+  ///   • `ERROR_INVALID_EMAIL` - If the [email] address is malformed.
+  ///   • `ERROR_WRONG_PASSWORD` - If the [password] is wrong.
+  ///   • `ERROR_USER_NOT_FOUND` - If there is no user corresponding to the given [email] address, or if the user has been deleted.
+  ///   • `ERROR_USER_DISABLED` - If the user has been disabled (for example, in the Firebase console)
+  ///   • `ERROR_TOO_MANY_REQUESTS` - If there was too many attempts to sign in as this user.
+  ///   • `ERROR_OPERATION_NOT_ALLOWED` - Indicates that Email & Password accounts are not enabled.
+  ///
   Future<FirebaseUser> signInWithEmailAndPassword({
     @required String email,
     @required String password,
@@ -259,35 +434,24 @@ class FirebaseAuth {
     return currentUser;
   }
 
-  Future<FirebaseUser> signInWithFacebook(
-      {@required String accessToken}) async {
-    assert(accessToken != null);
-    final Map<dynamic, dynamic> data =
-        await channel.invokeMethod('signInWithFacebook', <String, String>{
-      'accessToken': accessToken,
-    });
-    final FirebaseUser currentUser = new FirebaseUser._(data);
-    return currentUser;
-  }
-
-  /// Signs in with a Twitter account using the specified credentials.
+  /// Tries to sign in a user with the given Google [idToken] and [accessToken].
   ///
-  /// The returned future completes with the signed-in user or a [PlatformException], if sign in failed.
-  Future<FirebaseUser> signInWithTwitter({
-    @required String authToken,
-    @required String authTokenSecret,
-  }) async {
-    assert(authToken != null);
-    assert(authTokenSecret != null);
-    final Map<dynamic, dynamic> data =
-        await channel.invokeMethod('signInWithTwitter', <String, String>{
-      'authToken': authToken,
-      'authTokenSecret': authTokenSecret,
-    });
-    final FirebaseUser currentUser = new FirebaseUser._(data);
-    return currentUser;
-  }
-
+  /// If successful, it also signs the user in into the app and updates
+  /// the [onAuthStateChanged] stream.
+  ///
+  /// If the user doesn't have an account already, one will be created automatically.
+  ///
+  /// **Important**: You must enable Google accounts in the Auth section
+  /// of the Firebase console before being able to use them.
+  ///
+  /// Errors:
+  ///   • `ERROR_INVALID_CREDENTIAL` - If the [authToken] or [authTokenSecret] is malformed or has expired.
+  ///   • `ERROR_USER_DISABLED` - If the user has been disabled (for example, in the Firebase console)
+  ///   • `ERROR_ACCOUNT_EXISTS_WITH_DIFFERENT_CREDENTIAL` - If there already exists an account with the email address asserted by Google.
+  ///       Resolve this case by calling [fetchSignInMethodsForEmail] and then asking the user to sign in using one of them.
+  ///       This error will only be thrown if the "One account per email address" setting is enabled in the Firebase console (recommended).
+  ///   • `ERROR_OPERATION_NOT_ALLOWED` - Indicates that Google accounts are not enabled.
+  ///
   Future<FirebaseUser> signInWithGoogle({
     @required String idToken,
     @required String accessToken,
@@ -305,6 +469,83 @@ class FirebaseAuth {
     return currentUser;
   }
 
+  /// Tries to sign in a user with the given Facebook [accessToken].
+  ///
+  /// If successful, it also signs the user in into the app and updates
+  /// the [onAuthStateChanged] stream.
+  ///
+  /// If the user doesn't have an account already, one will be created automatically.
+  ///
+  /// **Important**: You must enable Facebook accounts in the Auth section
+  /// of the Firebase console before being able to use them.
+  ///
+  /// Errors:
+  ///   • `ERROR_INVALID_CREDENTIAL` - If the [accessToken] is malformed or has expired.
+  ///   • `ERROR_USER_DISABLED` - If the user has been disabled (for example, in the Firebase console)
+  ///   • `ERROR_ACCOUNT_EXISTS_WITH_DIFFERENT_CREDENTIAL` - If there already exists an account with the email address asserted by Facebook.
+  ///       Resolve this case by calling [fetchSignInMethodsForEmail] and then asking the user to sign in using one of them.
+  ///       This error will only be thrown if the "One account per email address" setting is enabled in the Firebase console (recommended).
+  ///   • `ERROR_OPERATION_NOT_ALLOWED` - Indicates that Facebook accounts are not enabled.
+  ///
+  Future<FirebaseUser> signInWithFacebook(
+      {@required String accessToken}) async {
+    assert(accessToken != null);
+    final Map<dynamic, dynamic> data =
+        await channel.invokeMethod('signInWithFacebook', <String, String>{
+      'accessToken': accessToken,
+    });
+    final FirebaseUser currentUser = new FirebaseUser._(data);
+    return currentUser;
+  }
+
+  /// Tries to sign in a user with the given Twitter [authToken] and [authTokenSecret].
+  ///
+  /// If successful, it also signs the user in into the app and updates
+  /// the [onAuthStateChanged] stream.
+  ///
+  /// If the user doesn't have an account already, one will be created automatically.
+  ///
+  /// **Important**: You must enable Twitter accounts in the Auth section
+  /// of the Firebase console before being able to use them.
+  ///
+  /// Errors:
+  ///   • `ERROR_INVALID_CREDENTIAL` - If the [authToken] or [authTokenSecret] is malformed or has expired.
+  ///   • `ERROR_USER_DISABLED` - If the user has been disabled (for example, in the Firebase console)
+  ///   • `ERROR_ACCOUNT_EXISTS_WITH_DIFFERENT_CREDENTIAL` - If there already exists an account with the email address asserted by Twitter.
+  ///       Resolve this case by calling [fetchSignInMethodsForEmail] and then asking the user to sign in using one of them.
+  ///       This error will only be thrown if the "One account per email address" setting is enabled in the Firebase console (recommended).
+  ///   • `ERROR_OPERATION_NOT_ALLOWED` - Indicates that Twitter accounts are not enabled.
+  ///
+  Future<FirebaseUser> signInWithTwitter({
+    @required String authToken,
+    @required String authTokenSecret,
+  }) async {
+    assert(authToken != null);
+    assert(authTokenSecret != null);
+    final Map<dynamic, dynamic> data =
+        await channel.invokeMethod('signInWithTwitter', <String, String>{
+      'authToken': authToken,
+      'authTokenSecret': authTokenSecret,
+    });
+    final FirebaseUser currentUser = new FirebaseUser._(data);
+    return currentUser;
+  }
+
+  /// Tries to sign in a user with the given Phone [verificationId] and [smsCode].
+  ///
+  /// If successful, it also signs the user in into the app and updates
+  /// the [onAuthStateChanged] stream.
+  ///
+  /// If the user doesn't have an account already, one will be created automatically.
+  ///
+  /// **Important**: You must enable Phone accounts in the Auth section
+  /// of the Firebase console before being able to use them.
+  ///
+  /// Errors:
+  ///   • `ERROR_INVALID_CREDENTIAL` - If the [verificationId] or [smsCode] is malformed or has expired.
+  ///   • `ERROR_USER_DISABLED` - If the user has been disabled (for example, in the Firebase console)
+  ///   • `ERROR_OPERATION_NOT_ALLOWED` - Indicates that Phone accounts are not enabled.
+  ///
   Future<FirebaseUser> signInWithPhoneNumber({
     @required String verificationId,
     @required String smsCode,
@@ -320,6 +561,49 @@ class FirebaseAuth {
     return currentUser;
   }
 
+  /// Starts the phone number verification process for the given phone number.
+  ///
+  /// Either sends an SMS with a 6 digit code to the phone number specified,
+  /// or sign's the user in and [verificationCompleted] is called.
+  ///
+  /// No duplicated SMS will be sent out upon re-entry (before timeout).
+  ///
+  /// Make sure to test all scenarios below:
+  ///   • You directly get logged in if Google Play Services verified the phone
+  ///     number instantly or helped you auto-retrieve the verification code.
+  ///   • Auto-retrieve verification code timed out.
+  ///   • Error cases when you receive [verificationFailed] callback.
+  ///
+  /// [phoneNumber] The phone number for the account the user is signing up
+  ///   for or signing into. Make sure to pass in a phone number with country
+  ///   code prefixed with plus sign ('+').
+  ///
+  /// [timeout] The maximum amount of time you are willing to wait for SMS
+  ///   auto-retrieval to be completed by the library. Maximum allowed value
+  ///   is 2 minutes. Use 0 to disable SMS-auto-retrieval. Setting this to 0
+  ///   will also cause [codeAutoRetrievalTimeout] to be called immediately.
+  ///   If you specified a positive value less than 30 seconds, library will
+  ///   default to 30 seconds.
+  ///
+  /// [forceResendingToken] The [forceResendingToken] obtained from [codeSent]
+  ///   callback to force re-sending another verification SMS before the
+  ///   auto-retrieval timeout.
+  ///
+  /// [verificationCompleted] This callback must be implemented.
+  ///   It will trigger when an SMS is auto-retrieved or the phone number has
+  ///   been instantly verified. The callback will provide a [FirebaseUser].
+  ///
+  /// [verificationFailed] This callback must be implemented.
+  ///   Triggered when an error occurred during phone number verification.
+  ///
+  /// [codeSent] Optional callback.
+  ///   It will trigger when an SMS has been sent to the users phone,
+  ///   and will include a [verificationId] and [forceResendingToken].
+  ///
+  /// [codeAutoRetrievalTimeout] Optional callback.
+  ///   It will trigger when SMS auto-retrieval times out and provide a
+  ///   [verificationId].
+  ///
   Future<void> verifyPhoneNumber({
     @required String phoneNumber,
     @required Duration timeout,
@@ -348,6 +632,25 @@ class FirebaseAuth {
     await channel.invokeMethod('verifyPhoneNumber', params);
   }
 
+  /// Tries to sign in a user with a given Custom Token [token].
+  ///
+  /// If successful, it also signs the user in into the app and updates
+  /// the [onAuthStateChanged] stream.
+  ///
+  /// Use this method after you retrieve a Firebase Auth Custom Token from your server.
+  ///
+  /// If the user identified by the [uid] specified in the token doesn't
+  /// have an account already, one will be created automatically.
+  ///
+  /// Read how to use Custom Token authentication and the cases where it is
+  /// useful in [the guides](https://firebase.google.com/docs/auth/android/custom-auth).
+  ///
+  /// Errors:
+  ///   • `ERROR_INVALID_CUSTOM_TOKEN` - The custom token format is incorrect.
+  ///     Please check the documentation.
+  ///   • `ERROR_CUSTOM_TOKEN_MISMATCH` - Invalid configuration.
+  ///     Ensure your app's SHA1 is correct in the Firebase console.
+  ///
   Future<FirebaseUser> signInWithCustomToken({@required String token}) async {
     assert(token != null);
     final Map<dynamic, dynamic> data = await channel.invokeMethod(
@@ -360,11 +663,16 @@ class FirebaseAuth {
     return currentUser;
   }
 
+  /// Signs out the current user and clears it from the disk cache.
+  ///
+  /// If successful, it signs the user out of the app and updates
+  /// the [onAuthStateChanged] stream.
+  ///
   Future<void> signOut() async {
     return await channel.invokeMethod("signOut");
   }
 
-  /// Asynchronously gets current user, or `null` if there is none.
+  /// Returns the currently signed-in [FirebaseUser] or [null] if there is none.
   Future<FirebaseUser> currentUser() async {
     final Map<dynamic, dynamic> data =
         await channel.invokeMethod("currentUser");
@@ -373,12 +681,20 @@ class FirebaseAuth {
     return currentUser;
   }
 
-  /// Links email account with current user and returns [Future<FirebaseUser>]
-  /// basically current user with additional email information
+  /// Links the given [email] and [password] to the current user.
   ///
-  /// throws [PlatformException] when
-  /// 1. email address is already used
-  /// 2. wrong email and password provided
+  /// This allows the user to sign in to this account in the future with
+  /// the given [email] and [password].
+  ///
+  /// Errors:
+  ///   • `ERROR_WEAK_PASSWORD` - If the password is not strong enough.
+  ///   • `ERROR_INVALID_CREDENTIAL` - If the email address is malformed.
+  ///   • `ERROR_CREDENTIAL_ALREADY_IN_USE` - If the email is already in use by a different account.
+  ///   • `ERROR_USER_DISABLED` - If the user has been disabled (for example, in the Firebase console)
+  ///   • `ERROR_REQUIRES_RECENT_LOGIN` - If the user's last sign-in time does not meet the security threshold.
+  ///   • `ERROR_PROVIDER_ALREADY_LINKED` - If the current user already has an Email & Password linked.
+  ///   • `ERROR_OPERATION_NOT_ALLOWED` - Indicates that Email & Password accounts are not enabled.
+  ///
   Future<FirebaseUser> linkWithEmailAndPassword({
     @required String email,
     @required String password,
@@ -396,14 +712,19 @@ class FirebaseAuth {
     return currentUser;
   }
 
-  /// Links google account with current user and returns [Future<FirebaseUser>]
+  /// Links the Google Account to the current user using [idToken] and [accessToken].
   ///
-  /// throws [PlatformException] when
-  /// 1. No current user provided (user has not logged in)
-  /// 2. No google credentials were found for given [idToken] and [accessToken]
-  /// 3. Google account already linked with another [FirebaseUser]
-  /// Detailed documentation on possible error causes can be found in [Android docs](https://firebase.google.com/docs/reference/android/com/google/firebase/auth/FirebaseUser#exceptions_4) and [iOS docs](https://firebase.google.com/docs/reference/ios/firebaseauth/api/reference/Classes/FIRUser#/c:objc(cs)FIRUser(im)linkWithCredential:completion:)
-  /// TODO: Throw custom exceptions with error codes indicating cause of exception
+  /// This allows the user to sign in to this account in the future with
+  /// the given Google Account.
+  ///
+  /// Errors:
+  ///   • `ERROR_INVALID_CREDENTIAL` - If the [authToken] or [authTokenSecret] is malformed or has expired.
+  ///   • `ERROR_CREDENTIAL_ALREADY_IN_USE` - If the Google account is already in use by a different account.
+  ///   • `ERROR_USER_DISABLED` - If the user has been disabled (for example, in the Firebase console)
+  ///   • `ERROR_REQUIRES_RECENT_LOGIN` - If the user's last sign-in time does not meet the security threshold.
+  ///   • `ERROR_PROVIDER_ALREADY_LINKED` - If the current user already has a Google account linked.
+  ///   • `ERROR_OPERATION_NOT_ALLOWED` - Indicates that Google accounts are not enabled.
+  ///
   Future<FirebaseUser> linkWithGoogleCredential({
     @required String idToken,
     @required String accessToken,
@@ -421,6 +742,19 @@ class FirebaseAuth {
     return currentUser;
   }
 
+  /// Links the Facebook Account to the current user using [accessToken].
+  ///
+  /// This allows the user to sign in to this account in the future with
+  /// the given Facebook Account.
+  ///
+  /// Errors:
+  ///   • `ERROR_INVALID_CREDENTIAL` - If the [accessToken] is malformed or has expired.
+  ///   • `ERROR_CREDENTIAL_ALREADY_IN_USE` - If the Facebook account is already in use by a different account.
+  ///   • `ERROR_USER_DISABLED` - If the user has been disabled (for example, in the Firebase console)
+  ///   • `ERROR_REQUIRES_RECENT_LOGIN` - If the user's last sign-in time does not meet the security threshold.
+  ///   • `ERROR_PROVIDER_ALREADY_LINKED` - If the current user already has a Facebook account linked.
+  ///   • `ERROR_OPERATION_NOT_ALLOWED` - Indicates that Facebook accounts are not enabled.
+  ///
   Future<FirebaseUser> linkWithFacebookCredential({
     @required String accessToken,
   }) async {
@@ -429,6 +763,36 @@ class FirebaseAuth {
       'linkWithFacebookCredential',
       <String, String>{
         'accessToken': accessToken,
+      },
+    );
+    final FirebaseUser currentUser = new FirebaseUser._(data);
+    return currentUser;
+  }
+
+  /// Links the Twitter Account to the current user using [authToken] and [authTokenSecret].
+  ///
+  /// This allows the user to sign in to this account in the future with
+  /// the given Twitter Account.
+  ///
+  /// Errors:
+  ///   • `ERROR_INVALID_CREDENTIAL` - If the [authToken] or [authTokenSecret] is malformed or has expired.
+  ///   • `ERROR_CREDENTIAL_ALREADY_IN_USE` - If the Twitter account is already in use by a different account.
+  ///   • `ERROR_USER_DISABLED` - If the user has been disabled (for example, in the Firebase console)
+  ///   • `ERROR_REQUIRES_RECENT_LOGIN` - If the user's last sign-in time does not meet the security threshold.
+  ///   • `ERROR_PROVIDER_ALREADY_LINKED` - If the current user already has a Twitter account linked.
+  ///   • `ERROR_OPERATION_NOT_ALLOWED` - Indicates that Twitter accounts are not enabled.
+  ///
+  Future<FirebaseUser> linkWithTwitterCredential({
+    @required String authToken,
+    @required String authTokenSecret,
+  }) async {
+    assert(authToken != null);
+    assert(authTokenSecret != null);
+    final Map<dynamic, dynamic> data = await channel.invokeMethod(
+      'linkWithTwitterCredential',
+      <String, String>{
+        'authToken': authToken,
+        'authTokenSecret': authTokenSecret,
       },
     );
     final FirebaseUser currentUser = new FirebaseUser._(data);